+++ src/vectors.c
@@ -6,7 +6,7 @@
 /*   By: vkuikka <vkuikka@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/07 18:13:00 by vkuikka           #+#    #+#             */
-/*   Updated: 2021/02/09 19:35:10 by vkuikka          ###   ########.fr       */
+/*   Updated: 2021/02/09 19:36:59 by vkuikka          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -101,7 +101,7 @@ float		vec_angle(t_vec3 v1, t_vec3 v2)
 	return (acos(dot / len));
 }
 
-void		vec_mult(t_vec3 *res, int mult)
+void		vec_mult(t_vec3 *res, float mult)
 {
 	res->x = res->x * mult;
 	res->y = res->y * mult;
